feat: add grpc server and all grpc methods

+++ catalog_service/pkg/entity/topping.go
@@ -1,7 +1,7 @@
 package entity
 
 type Topping struct {
-	Id    int
+	Id    uint64
 	Name  string
 	Price string
 }
