+++ catalog_service/pkg/repository/product_repository.go
@@ -14,7 +14,7 @@ func NewProductRepository(db *sqlx.DB) *ProductRepository {
 	return &ProductRepository{db}
 }
 
-func (r *ProductRepository) GetProductByID(id int) (entity.Product, error) {
+func (r *ProductRepository) GetProductByID(id uint64) (entity.Product, error) {
 	var result entity.Product
 	err := r.db.QueryRowx("SELECT * FROM products WHERE id = $1 LIMIT 1", id).StructScan(&result)
 	return result, err
